+++ packages/pass-style/src/types.test-d.ts
@@ -1,0 +1,31 @@
+/* eslint-disable */
+import { expectType, expectNotType } from 'tsd';
+import { Far } from './make-far';
+import { passStyleOf } from './passStyleOf';
+import { makeTagged } from './makeTagged';
+import { CopyTagged, PassStyle } from './types';
+import { PASS_STYLE } from './passStyle-helpers';
+
+const remotable = Far('foo', {});
+
+const copyTagged = makeTagged('someTag', remotable);
+expectType<CopyTagged<'someTag', typeof remotable>>(copyTagged);
+
+const someUnknown: unknown = null;
+
+expectType<'undefined'>(passStyleOf(undefined));
+expectType<'string'>(passStyleOf('str'));
+expectType<'boolean'>(passStyleOf(true));
+expectType<'number'>(passStyleOf(1));
+expectType<'bigint'>(passStyleOf(1n));
+expectType<'symbol'>(passStyleOf(Symbol.for('foo')));
+expectType<'null'>(passStyleOf(null));
+expectType<'promise'>(passStyleOf(Promise.resolve()));
+expectType<'error'>(passStyleOf(new Error()));
+expectType<'tagged'>(passStyleOf(copyTagged));
+expectType<'copyArray'>(passStyleOf([]));
+expectType<'copyRecord'>(passStyleOf({}));
+// though the object is specifying a PASS_STYLE, it doesn't match the case for extracting it
+expectType<'copyRecord'>(passStyleOf({ [PASS_STYLE]: 'arbitrary' } as const));
+expectType<'remotable'>(passStyleOf(remotable));
+expectType<PassStyle>(passStyleOf(someUnknown));