+++ packages/ses/src/commons.js
@@ -133,6 +133,10 @@ export const { prototype: weakmapPrototype } = WeakMap;
 export const { prototype: weaksetPrototype } = WeakSet;
 export const { prototype: functionPrototype } = Function;
 export const { prototype: promisePrototype } = Promise;
+export const { prototype: generatorPrototype } = getPrototypeOf(
+  // eslint-disable-next-line no-empty-function, func-names
+  function* () {},
+);
 
 export const typedArrayPrototype = getPrototypeOf(Uint8Array.prototype);
 
@@ -191,6 +195,9 @@ export const stringEndsWith = uncurryThis(stringPrototype.endsWith);
 export const stringIncludes = uncurryThis(stringPrototype.includes);
 export const stringIndexOf = uncurryThis(stringPrototype.indexOf);
 export const stringMatch = uncurryThis(stringPrototype.match);
+export const generatorNext = uncurryThis(generatorPrototype.next);
+export const generatorThrow = uncurryThis(generatorPrototype.throw);
+
 /**
  * @type { &
  *   ((thisArg: string, searchValue: { [Symbol.replace](string: string, replaceValue: string): string; }, replaceValue: string) => string) &