+++ packages/ses/NEWS.md
@@ -1,5 +1,10 @@
 User-visible changes in SES:
 
+# Next release
+
+- Adds `importNowHook` to the `Compartment` options. The compartment will invoke the hook whenever it encounters a missing dependency while running `compartmentInstance.importNow(specifier)`, which cannot use an asynchronous `importHook`.
+- To support TypeScript v5.5, a CommonJS-specific type declaration file is now explicitly exported.
+
 # v1.3.0 (2024-02-22)
 
 - Now supports `Promise.any`, `AggegateError`, `error.errors`,